9796-refactor check workitem on calendared case to own file
