+++ web-client/integration-tests/docketClerkDocumentQCJourney.test.js
@@ -1,9 +1,8 @@
 import {
-  CASE_STATUS_TYPES,
   COUNTRY_TYPES,
   PARTY_TYPES,
 } from '../../shared/src/business/entities/EntityConstants';
-import { createISODateString } from '../../shared/src/business/utilities/DateHandler';
+import { checkWorkitemOnCalendaredCase } from './journey/checkWorkitemOnCalendaredCase';
 import { docketClerkAddsAndServesDocketEntryFromOrder } from './journey/docketClerkAddsAndServesDocketEntryFromOrder';
 import { docketClerkAssignWorkItemToSelf } from './journey/docketClerkAssignWorkItemToSelf';
 import { docketClerkCreatesATrialSession } from './journey/docketClerkCreatesATrialSession';
@@ -29,7 +28,7 @@ describe('Docket Clerk Document QC Journey', () => {
     cerebralTest.closeSocket();
   });
 
-  const trialLocation = `Pheonix, Arizona, ${Date.now()}`;
+  const trialLocation = `Phoenix, Arizona, ${Date.now()}`;
   const overrides = {
     maxCases: 3,
     preferredTrialCity: trialLocation,
@@ -86,21 +85,15 @@ describe('Docket Clerk Document QC Journey', () => {
   docketClerkViewsQCInProgress(cerebralTest, false);
   docketClerkViewsQCOutbox(cerebralTest, true);
 
-  it('should have a trial date and trial location on workItem associated with a calendared case', () => {
-    const workQueue = cerebralTest.getState('workQueue');
-    const calendaredWorkItem = workQueue.find(
-      workItem =>
-        workItem.docketEntry.docketEntryId === cerebralTest.docketEntryId,
-    );
-
-    const expectedTrialDate = createISODateString(
-      `${overrides.trialDate.year}-${overrides.trialDate.month}-${overrides.trialDate.day}`,
-    );
-
-    expect(calendaredWorkItem.caseStatus).toBe(CASE_STATUS_TYPES.calendared);
-    expect(calendaredWorkItem.trialDate).toEqual(expectedTrialDate);
-    expect(calendaredWorkItem.trialLocation).toEqual(trialLocation);
-  });
+  checkWorkitemOnCalendaredCase(
+    cerebralTest,
+    {
+      day: overrides.trialDate.day,
+      month: overrides.trialDate.month,
+      year: overrides.trialDate.year,
+    },
+    trialLocation,
+  );
 
   loginAs(cerebralTest, 'petitionsclerk@example.com');
   petitionsClerkServesElectronicCaseToIrs(cerebralTest);
